Change view model to helper
