+++ app/src/main/java/org/wikipedia/recurring/RecommendedReadingListTask.kt
@@ -1,7 +1,7 @@
 package org.wikipedia.recurring
 
+import org.wikipedia.readinglist.recommended.RecommendedReadingListHelper
 import org.wikipedia.readinglist.recommended.RecommendedReadingListUpdateFrequency
-import org.wikipedia.readinglist.recommended.RecommendedReadingListViewModel
 import org.wikipedia.settings.Prefs
 import java.time.LocalDate
 import java.util.Date
@@ -25,6 +25,6 @@ class RecommendedReadingListTask() : RecurringTask() {
     }
 
     override suspend fun run(lastRun: Date) {
-        RecommendedReadingListViewModel.generateRecommendedReadingList()
+        RecommendedReadingListHelper.generateRecommendedReadingList()
     }
 }
