+++ app/src/main/res/layout/fragment_reading_list.xml
@@ -79,4 +79,19 @@
 
     </org.wikipedia.views.WikiSwipeRefreshLayout>
 
+    <org.wikipedia.views.WikiErrorView
+        android:id="@+id/errorView"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:orientation="vertical"
+        android:visibility="gone"/>
+
+    <com.google.android.material.progressindicator.LinearProgressIndicator
+        android:id="@+id/progressBar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_gravity="top"
+        android:indeterminate="true" />
+
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
