+++ app/src/main/res/values/strings.xml
@@ -1903,9 +1903,9 @@
         <item quantity="one">• %d article</item>
         <item quantity="other">• %d articles</item>
     </plurals>
-    <string name="recommended_reading_list_page_description_daily">Your daily reading list. Learn about new topics,  picked just for you. Updates at midnight.</string>
+    <string name="recommended_reading_list_page_description_daily">Your daily reading list. Learn about new topics, picked just for you. Updates at midnight.</string>
     <string name="recommended_reading_list_page_description_weekly">Your weekly reading list. Learn about new topics, picked just for you. Updates at the beginning of the week.</string>
-    <string name="recommended_reading_list_page_description_monthly">Your monthly reading list. Learn about new topics,  picked just for you. Updates on the first of each month.</string>
+    <string name="recommended_reading_list_page_description_monthly">Your monthly reading list. Learn about new topics, picked just for you. Updates on the first of each month.</string>
     <string name="recommended_reading_list_page_overflow_customize">Customize</string>
     <string name="recommended_reading_list_page_overflow_about">About this feature</string>
     <string name="recommended_reading_list_page_snackbar">Discover will update with %1$d new articles each %2$s. Your recommendations are stored locally on your device.</string>
